backend/pyproject.toml
[project]
name = "expense-tracker-backend"
version = "0.1.0"
description = "FastAPI backend (in-memory) for expense tracker"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.5",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "bcrypt>=4.0.0",
  "email-validator>=2.0.0.post2",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.0",
  "anyio>=4.0.0",
  "pytest>=7.4.0",
  "pytest-asyncio>=0.21.0",
  "pytest-cov>=4.1.0",
  "pytest-mock>=3.11.0",
  "black>=23.0.0",
  "isort>=5.12.0",
]

[tool.uv]
package = true

# Black 配置
[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # 排除目录
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

# isort 配置（与 Black 兼容）
[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

# 项目脚本（这些是 Python 入口点，不是 shell 命令）
# 要运行格式化命令，请使用：uv run black . 或直接使用 black 命令

